simplify the structure
